[BUGFIX] Functional tests use outdated objects
The functional tests did not clear all objects from memory when
resetting the persistence manager state. This lead to different
representations for the same entity being present in memory, which in
turn lead to errors when trying to compare objects. This is now
mitigated by replacing the election instance, which in turn should
also replace all other object instances.