[FEATURE] Terminate old voting sessions
This adds a new check at the beginning of a voting session, if another
session is already running. If yes, these sessions are terminated.

The termination is done in a two-step process: First a message is sent
via the SysV IPC (inter-process communication) system; this should work
on all *NIX systems. After that the PID gets a SIGUSR1 signal so it
reads the message and terminates. The only problem is that the fgets()
call used to read the client input is a blocking call, so while the
server waits for client input, the server process won't be terminated.
However, the check if the ballot box is still available is done between
reading and processing the command, so the next command won't be
executed after the session is terminated.