[BUGFIX] Ballot box status uses old sessions
As there is no mechanism to remove stale sessions and sessions are also
not marked as ended when a client crashes or the network connection is
dropped, there might be stale sessions lurking around in the database.
This commit lets the box status listing cope with that by checking if a
session is still running before using it as the current session.

An additional step would be to clean up stale sessions, probably with a
different command.