[BUGFIX] Ballot box/session objects not updated regularly
As sessions might run for a long time, the object's state might change
in the database. Until now, this was not reflected in the code, which
used the same objects during the whole session. This change updates the
objects with each new command. Changes made to the objects will not be
persisted, but that shouldn't be a problem as the ballot box session
should not change the ballot box object – and the session object is not
changed.